make app name dynamic in docker-test

+++ docker_steps/docker-test.sh
@@ -6,5 +6,9 @@ docker rm -fv $container_name || true
 test_container_id=`docker run --name $container_name -d -w /app ***REMOVED******REMOVED***-calculator:develop /bin/bash -c "while true; do sleep 10; done"`
 docker cp test-requires $test_container_id:/tmp/test-requires
 docker exec -it "$test_container_id" pip3 install -r /tmp/test-requires
-docker exec -it "$test_container_id" pylint -E ***REMOVED******REMOVED***_calculator
+if [ -z $APP_NAME ]; then
+    echo Need APP_NAME variable
+    exit 1
+fi
+docker exec -it "$test_container_id" pylint -E $APP_NAME
 docker rm -fv $test_container_id
